chore(move): ln_bridge/commander => ln_cmd
