+++ src/ln_bridge/mod.rs
@@ -2,7 +2,7 @@ pub mod key;
 pub mod utils;
 pub mod channel_manager;
 pub mod log_printer;
-pub mod commander;
+// pub mod commander;
 pub mod settings;
 
 pub trait Restorable<R, T> {
